Doesn't work
The sample renamed to Balloon App 2 works, but when I copied all code
from sample to here, it doesn't.
