+++ app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -63,4 +62,3 @@
         android:layout_marginStart="@dimen/small_margin"
         android:textSize="@dimen/default_text_size" />
 </LinearLayout>
-
