+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">Balloon App</string>
+    <string name="app_name">Balloon App </string>
 
     <string name="latitude_label">Latitude</string>
     <string name="longitude_label">Longitude</string>
